requirements.txt
django 
djangorestframework
environs
django-cors-headers
dj-database-url
mysqlclient